gallium: Use a single draw to emulate primitive restart
The spec says that primitive restart does not affect gl_PrimitiveIDIn in a
geometry shader, but if we split up the draw the primitive ID will get reset on
restart. The fix is to use a single draw with an unrolled index buffer.

As a worked example, if we want to draw the following TRIANGLE_STRIPS

   0, 1, 2, 3, -1, 6, 7, 8

we instead draw the following TRIANGLES

   0, 1, 2, 1, 2, 3, 6, 7, 8

This correctly draws 3 triangles with primitive IDs 0, 1, 2 as intended.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>